[Major] Renaming master's NewSelect for merging
