+++ lib/src/select/Select.jsx
@@ -5,8 +5,8 @@ import { spaces } from "../common/variables.js";
 import BackgroundColorContext from "../BackgroundColorContext.js";
 import { v4 as uuidv4 } from "uuid";
 import { getMargin } from "../common/utils.js";
-
-const getNotOptionalErrorMessage = () => `This field is required. Please, enter a value.`;
+import { useLayoutEffect } from "react";
+import DxcCheckbox from "../checkbox/Checkbox";
 
 const selectIcons = {
   error: (
@@ -39,14 +39,7 @@ const selectIcons = {
     </svg>
   ),
   searchOff: (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      enable-background="new 0 0 24 24"
-      height="24px"
-      viewBox="0 0 24 24"
-      width="24px"
-      fill="currentColor"
-    >
+    <svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 0 24 24" width="24px" fill="currentColor">
       <g>
         <rect fill="none" height="24" width="24" />
       </g>
@@ -60,6 +53,8 @@ const selectIcons = {
   ),
 };
 
+const getNotOptionalErrorMessage = () => `This field is required. Please, enter a value.`;
+
 const DxcSelect = React.forwardRef(
   (
     {
@@ -72,7 +67,7 @@ const DxcSelect = React.forwardRef(
       disabled = false,
       optional = false,
       searchable = false,
-      // multiple = false,
+      multiple = false,
       onChange,
       onBlur,
       error = "",
@@ -83,103 +78,239 @@ const DxcSelect = React.forwardRef(
     ref
   ) => {
     const [selectId] = useState(`select-${uuidv4()}`);
-    const [innerValue, setInnerValue] = useState("");
+    const selectLabelId = `label-${selectId}`;
+    const [innerValue, setInnerValue] = useState(multiple ? [] : "");
     const [searchValue, setSearchValue] = useState("");
     const [isBackgroundValue, changeIsBackgroundValue] = useState(false);
 
     const [isOpen, changeIsOpen] = useState(false);
     const [isActiveOption, changeIsActiveOption] = useState(false);
+    const [isScrollable, changeIsScrollable] = useState(false);
     const [filteredOptions, setFilteredOptions] = useState([]);
-    const [visualFocusIndex, changeVisualFocusIndex] = useState(0);
+    const [visualFocusIndex, changeVisualFocusIndex] = useState(-1);
 
-    const selectInputRef = useRef(null);
+    const selectContainerRef = useRef(null);
+    const selectSearchInputRef = useRef(null);
+    const selectOptionsListRef = useRef(null);
 
     const colorsTheme = useTheme();
     const backgroundType = useContext(BackgroundColorContext);
 
+    const optionalEmptyOption = { label: placeholder, value: "" };
+
     const notOptionalCheck = (value) => value === "" && !optional;
+
+    const notOptionalMultipleCheck = () => (value ?? innerValue).length === 0 && !optional;
 
     const canBeOpenOptions = () => !disabled && options?.length > 0 && groupsHaveOptions();
     const groupsHaveOptions = () =>
-      options[0]?.hasOwnProperty("options")
+      options[0].hasOwnProperty("options")
         ? options[0].options
           ? options.some((groupOption) => groupOption.options.length > 0)
           : false
         : true;
     const filteredGroupsHaveOptions = () =>
-      filteredOptions[0]?.options ? filteredOptions.some((groupOption) => groupOption.options?.length > 0) : true;
+      filteredOptions?.[0].options ? filteredOptions.some((groupOption) => groupOption.options?.length > 0) : true;
+
     const openOptions = () => {
-      if (canBeOpenOptions()) {
+      if (!isOpen && canBeOpenOptions()) {
         searchable && changeIsBackgroundValue(true);
         changeIsOpen(true);
       }
     };
     const closeOptions = () => {
-      searchable && changeIsBackgroundValue(false);
-      changeIsOpen(false);
-      changeVisualFocusIndex(0);
+      if (isOpen) {
+        searchable && changeIsBackgroundValue(false);
+        changeIsOpen(false);
+        changeVisualFocusIndex(-1);
+      }
     };
 
-    const handleSelectChangeValue = (newValue) => {
-      value ?? setInnerValue(newValue);
+    const handleSelectChangeValue = (newOption) => {
+      if (multiple) {
+        let res;
+        if ((value ?? innerValue).includes(newOption.value))
+          value
+            ? (res = value.filter((optionVal) => optionVal !== newOption.value))
+            : setInnerValue((previous) => previous.filter((optionVal) => optionVal !== newOption.value));
+        else value ? (res = [...value, newOption.value]) : setInnerValue((previous) => [...previous, newOption.value]);
 
-      if (notOptionalCheck(newValue)) onChange?.({ value: newValue, error: getNotOptionalErrorMessage() });
-      else onChange?.({ value: newValue, error: null });
+        if (notOptionalMultipleCheck(newOption.value))
+          onChange?.({ value: res ?? innerValue, error: getNotOptionalErrorMessage() });
+        else onChange?.({ value: res ?? innerValue, error: null });
+      } else {
+        value ?? setInnerValue(newOption.value);
+
+        if (notOptionalCheck(newOption.value))
+          onChange?.({ value: newOption.value, error: getNotOptionalErrorMessage() });
+        else onChange?.({ value: newOption.value, error: null });
+      }
     };
     const handleSelectOnClick = () => {
+      changeVisualFocusIndex(0);
       isOpen ? closeOptions() : openOptions();
-      searchable && selectInputRef.current.focus();
+      searchable && selectSearchInputRef.current.focus();
     };
     const handleSelectOnFocus = () => {
-      searchable && selectInputRef.current.focus();
+      searchable && selectSearchInputRef.current.focus();
     };
     const handleSelectOnBlur = (event) => {
       // focus leaves container (outside, not to childs)
       if (!event.currentTarget.contains(event.relatedTarget)) {
-        setSearchValue("");
         closeOptions();
+        setSearchValue("");
 
         if (notOptionalCheck(value ?? innerValue))
           onBlur?.({ value: value ?? innerValue, error: getNotOptionalErrorMessage() });
         else onBlur?.({ value: value ?? innerValue, error: null });
       }
     };
+    const handleSelectOnKeyDown = (event) => {
+      switch (event.keyCode) {
+        case 40: // Arrow Down
+          event.preventDefault();
+          changeVisualFocusIndex((visualFocusIndex) => {
+            if (visualFocusIndex < lastOptionIndex) return visualFocusIndex + 1;
+            else if (visualFocusIndex === lastOptionIndex) return 0;
+          });
+          openOptions();
+          changeIsActiveOption(false);
+          isOpen && changeIsScrollable(true);
+          break;
+        case 38: // Arrow Up
+          event.preventDefault();
+          changeVisualFocusIndex((visualFocusIndex) =>
+            visualFocusIndex === 0 || visualFocusIndex === -1 ? lastOptionIndex : visualFocusIndex - 1
+          );
+          openOptions();
+          changeIsActiveOption(false);
+          isOpen && changeIsScrollable(true);
+          break;
+        case 27: // Esc
+          event.preventDefault();
+          closeOptions();
+          setSearchValue("");
+          break;
+        case 13: // Enter
+          if (isOpen) {
+            let accLength = optional && !multiple ? 1 : 0;
+            if (searchable) {
+              if (filteredOptions.length > 0) {
+                if (optional && !multiple && visualFocusIndex === 0 && filteredGroupsHaveOptions())
+                  handleSelectChangeValue(optionalEmptyOption);
+                else
+                  filteredOptions[0].options
+                    ? filteredGroupsHaveOptions() &&
+                      filteredOptions.some((groupOption) => {
+                        const groupLength = accLength + groupOption.options.length;
+                        groupLength > visualFocusIndex &&
+                          handleSelectChangeValue(groupOption.options[visualFocusIndex - accLength]);
+                        accLength = groupLength;
+                        return groupLength > visualFocusIndex;
+                      })
+                    : handleSelectChangeValue(filteredOptions[visualFocusIndex - accLength]);
+              }
+            } else {
+              if (optional && !multiple && visualFocusIndex === 0) handleSelectChangeValue(optionalEmptyOption);
+              else
+                options[0].options
+                  ? options.some((groupOption) => {
+                      const groupLength = accLength + groupOption.options.length;
+                      groupLength > visualFocusIndex &&
+                        handleSelectChangeValue(groupOption.options[visualFocusIndex - accLength]);
+                      accLength = groupLength;
+                      return groupLength > visualFocusIndex;
+                    })
+                  : handleSelectChangeValue(options[visualFocusIndex - accLength]);
+            }
+            !multiple && closeOptions();
+            setSearchValue("");
+          }
+          break;
+      }
+    };
 
     const handleSearchIOnChange = (event) => {
-      openOptions();
       setSearchValue(event.target.value);
+      !isOpen && changeVisualFocusIndex(0);
+      openOptions();
     };
 
     const handleClearActionOnClick = (event) => {
-      event.stopPropagation(); // not bubbling, clear not executes select event onClick
+      event.stopPropagation();
       setSearchValue("");
     };
 
-    const getSingleSelectedOption = () => {
-      const val = value ?? innerValue;
-      let selectedOption;
+    const handleClearOptionsActionOnClick = (event) => {
+      event.stopPropagation();
+      value ?? setInnerValue([]);
+      onChange?.({ value: [], error: getNotOptionalErrorMessage() });
+      selectContainerRef.current.focus();
+    };
 
-      if (options?.length > 0) {
-        options.forEach((option) => {
-          if (option.options) {
-            option.options.forEach((singleOption) => {
-              if (singleOption.value === val) selectedOption = singleOption;
-            });
-          } else if (option.value === val) selectedOption = option;
-        });
+    const getLastOptionIndex = () => {
+      let last = 0;
+      const reducer = (acc, current) => acc + current.options?.length;
+
+      if (searchable && filteredOptions.length > 0)
+        filteredOptions[0].options
+          ? (last = filteredOptions.reduce(reducer, 0) - 1)
+          : (last = filteredOptions.length - 1);
+      else if (options?.length > 0)
+        options[0].options ? (last = options.reduce(reducer, 0) - 1) : (last = options.length - 1);
+
+      return optional && !multiple ? last + 1 : last;
+    };
+    const lastOptionIndex = useMemo(
+      () => getLastOptionIndex(),
+      [searchable, optional, multiple, searchable ? filteredOptions : options]
+    );
+
+    const getSelectedOption = () => {
+      const val = value ?? innerValue;
+      let selectedOption = multiple ? [] : "";
+
+      if (multiple) {
+        if (options?.length > 0) {
+          options.forEach((option) => {
+            if (option.options) {
+              option.options.forEach((singleOption) => {
+                if (val.includes(singleOption.value)) selectedOption.push(singleOption);
+              });
+            } else if (val.includes(option.value)) selectedOption.push(option);
+          });
+        }
+      } else {
+        if (options?.length > 0) {
+          options.forEach((option) => {
+            if (option.options) {
+              option.options.forEach((singleOption) => {
+                if (singleOption.value === val) selectedOption = singleOption;
+              });
+            } else if (option.value === val) selectedOption = option;
+          });
+        }
       }
 
       return selectedOption;
     };
-    const selectedOption = useMemo(() => getSingleSelectedOption(), [value ?? innerValue]);
+    const selectedOption = useMemo(() => getSelectedOption(), [options, multiple, value ?? innerValue]);
+
+    useLayoutEffect(() => {
+      if (isScrollable) {
+        const visualFocusedOptionEl =
+          selectOptionsListRef?.current?.querySelectorAll("[role='option']")[visualFocusIndex];
+        visualFocusedOptionEl?.scrollIntoView({ block: "nearest", inline: "start" });
+        return changeIsScrollable(false);
+      }
+    }, [isScrollable]);
 
     useEffect(() => {
       if (searchable && options?.length > 0) {
-        changeVisualFocusIndex(0);
         if (options[0].options) {
           setFilteredOptions(
             options.map((optionGroup) => {
-              const group = JSON.parse(JSON.stringify(optionGroup)); // anything better?
+              const group = JSON.parse(JSON.stringify(optionGroup)); // circular issue
               group.options = group.options.filter((option) =>
                 option.label.toUpperCase().includes(searchValue.toUpperCase())
               );
@@ -193,36 +324,29 @@ const DxcSelect = React.forwardRef(
       }
     }, [options, searchable, searchValue]);
 
-    const lastIndex = () => {
-      let last = 0;
-      const reducer = (acc, current) => acc + current.options.length;
-
-      if (searchable && filteredOptions.length > 0)
-        filteredOptions[0].options
-          ? (last = filteredOptions.reduce(reducer, 0) - 1)
-          : (last = filteredOptions.length - 1);
-      else options[0]?.options ? (last = options.reduce(reducer, 0) - 1) : (last = options.length - 1);
-
-      return last;
-    };
-    const lastIndexValue = useMemo(() => lastIndex(), [searchable, searchable ? filteredOptions : options]);
+    useEffect(() => {
+      visualFocusIndex > lastOptionIndex && changeVisualFocusIndex(0);
+    }, [filteredOptions]);
 
     const Option = ({ option, index, isGroupedOption = false }) => {
-      const isSelected = (value ?? innerValue) === option.value;
-      const isLastOption = index === (optional ? lastIndexValue + 1 : lastIndexValue);
+      const isSelected = multiple
+        ? (value ?? innerValue).includes(option.value)
+        : (value ?? innerValue) === option.value;
+      const isLastOption = index === lastOptionIndex;
 
       return (
         <OptionItem
           onMouseDown={(event) => {
-            event.button === 0 && changeIsActiveOption(true); // left button only
+            // left mouse button only
+            event.button === 0 && changeIsActiveOption(true);
           }}
           onMouseUp={(event) => {
             if (event.button === 0 && isActiveOption) {
-              // left button only
-              handleSelectChangeValue(option.value);
+              // left mouse button only
+              handleSelectChangeValue(option);
+              !multiple && closeOptions();
               setSearchValue("");
               changeIsActiveOption(false);
-              closeOptions(); // if not multiple
             }
           }}
           onMouseEnter={() => {
@@ -234,25 +358,34 @@ const DxcSelect = React.forwardRef(
           visualFocused={visualFocusIndex === index}
           active={visualFocusIndex === index && isActiveOption}
           selected={isSelected}
+          aria-selected={isSelected && "true"}
+          role="option"
         >
           <StyledOption
             visualFocused={visualFocusIndex === index}
             active={visualFocusIndex === index && isActiveOption}
             selected={isSelected}
             last={isLastOption}
+            grouped={isGroupedOption}
+            multiple={multiple}
           >
-            <OptionContent grouped={isGroupedOption}>
-              {option.icon && <OptionIcon>{option.icon}</OptionIcon>}
+            {multiple && <DxcCheckbox tabIndex={-1} checked={isSelected} />}
+            {option.icon && (
+              <OptionIcon>
+                {typeof option.icon === "string" ? <OptionIconImg src={option.icon}></OptionIconImg> : option.icon}
+              </OptionIcon>
+            )}
+            <OptionContent grouped={isGroupedOption} hasIcon={option.icon} multiple={multiple}>
               <OptionLabel>{option.label}</OptionLabel>
+              {!multiple && isSelected && <SelectedIcon>{selectIcons.selected}</SelectedIcon>}
             </OptionContent>
-            {isSelected && <SelectedIcon>{selectIcons.selected}</SelectedIcon>}
           </StyledOption>
         </OptionItem>
       );
     };
 
-    let global_index = optional /*&& !multiple*/ ? 0 : -1; // placeholder becomes an option and its always index = 0
-    const mapOptionFunc = (option, index) => {
+    let global_index = optional && !multiple ? 0 : -1; // index for options (not groups), starting from 0 to options.length -1
+    const mapOptionFunc = (option) => {
       if (option.options) {
         return (
           <>
@@ -263,50 +396,91 @@ const DxcSelect = React.forwardRef(
             })}
           </>
         );
-      } else return <Option option={option} index={optional /*&& !multiple*/ ? index + 1 : index} />;
+      } else {
+        global_index++;
+        return <Option option={option} index={global_index} />;
+      }
     };
 
     return (
       <ThemeProvider theme={colorsTheme.select}>
         <DxcSelectContainer margin={margin} size={size} ref={ref}>
-          <Label htmlFor={selectId} disabled={disabled} backgroundType={backgroundType}>
+          <Label
+            id={selectLabelId}
+            disabled={disabled}
+            backgroundType={backgroundType}
+            onClick={() => {
+              selectContainerRef.current.focus();
+            }}
+          >
             {label} {optional && <OptionalLabel>(Optional)</OptionalLabel>}
           </Label>
           <HelperText disabled={disabled} backgroundType={backgroundType}>
             {helperText}
           </HelperText>
           <SelectContainer
+            id={selectId}
             backgroundType={backgroundType}
             disabled={disabled}
             error={error}
             onBlur={handleSelectOnBlur}
             onClick={handleSelectOnClick}
             onFocus={handleSelectOnFocus}
+            onKeyDown={handleSelectOnKeyDown}
+            ref={selectContainerRef}
             tabIndex={tabIndex}
+            aria-labelledby={selectLabelId}
           >
+            {multiple && selectedOption.length > 0 && (
+              <SelectionIndicator>
+                <SelectionValue>{selectedOption.length} </SelectionValue>
+                <ClearOptionsAction
+                  onClick={handleClearOptionsActionOnClick}
+                  backgroundType={backgroundType}
+                  tabIndex={-1}
+                  title="Clear selected options"
+                  aria-label="Clear selected options"
+                >
+                  {selectIcons.clear}
+                </ClearOptionsAction>
+              </SelectionIndicator>
+            )}
             <SearchableValueContainer>
+              <ValueInput name={name} value={value ?? innerValue} readOnly />
               {searchable && (
                 <SearchInput
-                  id={selectId}
-                  name={name}
                   value={searchValue}
                   disabled={disabled}
                   onChange={handleSearchIOnChange}
-                  ref={selectInputRef}
+                  ref={selectSearchInputRef}
                   autoComplete="off"
                   autoCorrect="off"
                 ></SearchInput>
               )}
-              {(!searchable || searchValue === "") && (
-                <SelectedOption disabled={disabled} placeholder={!(value ?? innerValue) || isBackgroundValue}>
-                  {selectedOption?.icon && <OptionIcon selected={true}>{selectedOption.icon}</OptionIcon>}
-                  <OptionLabel>{selectedOption?.label ?? placeholder}</OptionLabel>
-                </SelectedOption>
-              )}
+              {(!searchable || searchValue === "") &&
+                (multiple ? (
+                  <SelectedOption
+                    disabled={disabled}
+                    backgroundValue={(value ?? innerValue).length === 0 || isBackgroundValue}
+                  >
+                    <OptionLabel>{selectedOption.map((option) => option.label).join(", ")}</OptionLabel>
+                    {selectedOption.length === 0 && placeholder}
+                  </SelectedOption>
+                ) : (
+                  <SelectedOption disabled={disabled} backgroundValue={!(value ?? innerValue) || isBackgroundValue}>
+                    <OptionLabel>{selectedOption?.label ?? placeholder}</OptionLabel>
+                  </SelectedOption>
+                ))}
             </SearchableValueContainer>
             {!disabled && error && <ErrorIcon backgroundType={backgroundType}>{selectIcons.error}</ErrorIcon>}
             {searchable && searchValue.length > 0 && (
-              <ClearAction onClick={handleClearActionOnClick} backgroundType={backgroundType} tabIndex={-1}>
+              <ClearAction
+                onClick={handleClearActionOnClick}
+                backgroundType={backgroundType}
+                tabIndex={-1}
+                title="Clear search text"
+                aria-label="Clear search text"
+              >
                 {selectIcons.clear}
               </ClearAction>
             )}
@@ -321,6 +495,9 @@ const DxcSelect = React.forwardRef(
                 onMouseDown={(event) => {
                   event.preventDefault();
                 }}
+                ref={selectOptionsListRef}
+                role="listbox"
+                aria-label={label}
               >
                 {searchable && (filteredOptions.length === 0 || !filteredGroupsHaveOptions()) ? (
                   <OptionsSystemMessage>
@@ -328,7 +505,7 @@ const DxcSelect = React.forwardRef(
                     No matches found
                   </OptionsSystemMessage>
                 ) : (
-                  optional && /*!multiple &&*/ <Option option={{ label: placeholder, value: "" }} index={0} />
+                  optional && !multiple && <Option option={optionalEmptyOption} index={0} />
                 )}
                 {searchable ? filteredOptions.map(mapOptionFunc) : options.map(mapOptionFunc)}
               </OptionsList>
@@ -369,7 +546,7 @@ const DxcSelectContainer = styled.div`
     props.margin && typeof props.margin === "object" && props.margin.left ? spaces[props.margin.left] : ""};
 `;
 
-const Label = styled.label`
+const Label = styled.span`
   color: ${(props) =>
     props.disabled
       ? props.backgroundType === "dark"
@@ -384,6 +561,7 @@ const Label = styled.label`
   font-style: ${(props) => props.theme.labelFontStyle};
   font-weight: ${(props) => props.theme.labelFontWeight};
   line-height: 1.715em;
+  cursor: default;
 `;
 
 const OptionalLabel = styled.span`
@@ -471,6 +649,58 @@ const SelectContainer = styled.div`
     `};
 `;
 
+const SelectionIndicator = styled.span`
+  display: flex;
+  border: 1px solid ${(props) => props.theme.selectionIndicatorBorderColor};
+  border-radius: 2px;
+  width: 48px;
+`;
+
+const ClearOptionsAction = styled.button`
+  display: flex;
+  flex-wrap: wrap;
+  align-content: center;
+  height: 24px;
+  width: 24px;
+  font-size: 1rem;
+  font-family: ${(props) => props.theme.fontFamily};
+  border: 1px solid transparent;
+  padding: 3px;
+  ${(props) => (props.disabled ? `cursor: not-allowed;` : `cursor: pointer;`)}
+  background-color: ${(props) => props.theme.enabledSelectionIndicatorActionBackgroundColor};
+  color: ${(props) => props.theme.enabledSelectionIndicatorActionIconColor};
+
+  ${(props) =>
+    !props.disabled &&
+    `
+      &:hover {
+        background-color: ${props.theme.hoverSelectionIndicatorActionBackgroundColor};
+        color: ${props.theme.hoverSelectionIndicatorActionIconColor};
+      }
+      &:active {
+        background-color: ${props.theme.activeSelectionIndicatorActionBackgroundColor};
+        color: ${props.theme.activeSelectionIndicatorActionIconColor};
+      }
+    `}
+
+  svg {
+    line-height: 18px;
+  }
+`;
+
+const SelectionValue = styled.span`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  user-select: none;
+  background-color: ${(props) => props.theme.selectionIndicatorBackgroundColor};
+  border-right: 1px solid ${(props) => props.theme.selectionIndicatorBorderColor};
+  font-family: ${(props) => props.theme.fontFamily};
+  font-size: 11px;
+  ${(props) => (props.disabled ? `cursor: not-allowed;` : `cursor: default;`)}
+`;
+
 const SearchableValueContainer = styled.div`
   display: grid;
   width: 100%;
@@ -486,7 +716,7 @@ const SelectedOption = styled.span`
   overflow: hidden;
 
   color: ${(props) => {
-    if (props.placeholder)
+    if (props.backgroundValue)
       return props.disabled
         ? props.backgroundType === "dark"
           ? props.theme.disabledPlaceholderFontColorOnDark
@@ -509,6 +739,10 @@ const SelectedOption = styled.span`
   font-style: ${(props) => props.theme.valueFontStyle};
   font-weight: ${(props) => props.theme.valueFontWeight};
   line-height: 1.5em;
+`;
+
+const ValueInput = styled.input`
+  display: none;
 `;
 
 const SearchInput = styled.input`
@@ -602,7 +836,7 @@ const ClearAction = styled.button`
   font-size: 1rem;
   font-family: ${(props) => props.theme.fontFamily};
   border: 1px solid transparent;
-  border-radius: 4px;
+  border-radius: 2px;
   padding: 3px;
   margin-left: calc(1rem * 0.25);
   ${(props) => (props.disabled ? `cursor: not-allowed;` : `cursor: pointer;`)}
@@ -720,9 +954,9 @@ const OptionItem = styled.li`
 
 const StyledOption = styled.span`
   display: flex;
-  justify-content: space-between;
-  padding: 4px 8px 3px 8px;
+  padding: 4px 8px 3px 0;
   min-height: 24px;
+  ${(props) => props.grouped && props.multiple && `padding-left: 16px;`}
   ${(props) =>
     props.last
       ? `border-bottom: 1px solid transparent`
@@ -731,8 +965,11 @@ const StyledOption = styled.span`
 
 const OptionContent = styled.span`
   display: flex;
+  justify-content: space-between;
+  width: 100%;
   overflow: hidden;
-  ${(props) => props.grouped && `padding-left: 8px;`}
+  padding-left: 8px;
+  ${(props) => props.grouped && !props.multiple && !props.hasIcon && `padding-left: 16px;`}
 `;
 
 const OptionIcon = styled.span`
@@ -741,7 +978,12 @@ const OptionIcon = styled.span`
   align-content: center;
   height: 24px;
   width: 24px;
-  margin-right: ${(props) => (props.selected ? "4px" : "8px")};
+  margin-left: 8px;
+`;
+
+const OptionIconImg = styled.img`
+  width: 16px;
+  height: 16px;
 `;
 
 const OptionLabel = styled.span`
